Condensed/ re-done Focuses with button for Full Screen
Removed ID 9602 when clicking down and therefore reduced the code down to a more simple solution. "Click for Full Screen" label has been changed to a button. Clicking on up and down now shows the plot as it originally was. 

For TV shows, the episode thumbnail is only shown when Play button is displayed.